refactor(server): modularize main.py into core, routers, and services (closes #15)
